Joomla 1.7.0

+++ administrator/components/com_templates/models/style.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @version		$Id: style.php 21032 2011-03-29 16:38:31Z dextercowley $
+ * @version		$Id: style.php 21650 2011-06-23 05:29:17Z chdemko $
  * @package		Joomla.Administrator
  * @subpackage	com_templates
  * @copyright	Copyright (C) 2005 - 2011 Open Source Matters, Inc. All rights reserved.
@@ -80,6 +80,11 @@ class TemplatesModelStyle extends JModelAdmin
 				if (!$user->authorise('core.delete', 'com_templates')) {
 					throw new Exception(JText::_('JERROR_CORE_DELETE_NOT_PERMITTED'));
 				}
+				// You should not delete a default style
+				if ($table->home != '0'){
+					JError::raiseWarning(SOME_ERROR_NUMBER,Jtext::_('COM_TEMPLATES_STYLE_CANNOT_DELETE_DEFAULT_STYLE'));
+					return false;
+				}
 
 				if (!$table->delete($pk)) {
 					$this->setError($table->getError());
@@ -91,7 +96,7 @@ class TemplatesModelStyle extends JModelAdmin
 				return false;
 			}
 		}
-		
+
 		// Clean cache
 		$this->cleanCache();
 
@@ -145,7 +150,7 @@ class TemplatesModelStyle extends JModelAdmin
 				throw new Exception($table->getError());
 			}
 		}
-		
+
 		// Clean cache
 		$this->cleanCache();
 
@@ -250,7 +255,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 			// Convert the params field to an array.
 			$registry = new JRegistry;
-			$registry->loadJSON($table->params);
+			$registry->loadString($table->params);
 			$this->_cache[$pk]->params = $registry->toArray();
 
 			// Get the template XML.
@@ -350,6 +355,13 @@ class TemplatesModelStyle extends JModelAdmin
 	 */
 	public function save($data)
 	{
+		// Detect disabled extension
+		$extension = JTable::getInstance('Extension');
+		if ($extension->load(array('enabled' => 0, 'type' => 'template', 'element' => $data['template'], 'client_id' => $data['client_id']))) {
+			$this->setError(JText::_('COM_TEMPLATES_ERROR_SAVE_DISABLED_TEMPLATE'));
+			return false;
+		}
+
 		// Initialise variables;
 		$dispatcher = JDispatcher::getInstance();
 		$table		= $this->getTable();
@@ -369,7 +381,7 @@ class TemplatesModelStyle extends JModelAdmin
 		$data['home'] = 0;
 		$data['assigned'] ='';
 		}
-		
+
 		// Bind the data.
 		if (!$table->bind($data)) {
 			$this->setError($table->getError());
@@ -470,26 +482,23 @@ class TemplatesModelStyle extends JModelAdmin
 			throw new Exception(JText::_('JLIB_APPLICATION_ERROR_EDITSTATE_NOT_PERMITTED'));
 		}
 
-		// Lookup the client_id.
-		$db->setQuery(
-			'SELECT client_id' .
-			' FROM #__template_styles' .
-			' WHERE id = '.(int) $id
-		);
-		$clientId = $db->loadResult();
-
-		if ($error = $db->getErrorMsg()) {
-			throw new Exception($error);
-		}
-		else if (!is_numeric($clientId)) {
+		$style = JTable::getInstance('Style','TemplatesTable');
+		if (!$style->load((int)$id)) {
 			throw new Exception(JText::_('COM_TEMPLATES_ERROR_STYLE_NOT_FOUND'));
 		}
+
+		// Detect disabled extension
+		$extension = JTable::getInstance('Extension');
+		if ($extension->load(array('enabled' => 0, 'type' => 'template', 'element' => $style->template, 'client_id' => $style->client_id))) {
+			throw new Exception(JText::_('COM_TEMPLATES_ERROR_SAVE_DISABLED_TEMPLATE'));
+		}
+
 
 		// Reset the home fields for the client_id.
 		$db->setQuery(
 			'UPDATE #__template_styles' .
 			' SET home = \'0\'' .
-			' WHERE client_id = '.(int) $clientId .
+			' WHERE client_id = '.(int) $style->client_id .
 			' AND home = \'1\''
 		);
 
@@ -510,7 +519,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 		// Clean the cache.
 		$this->cleanCache();
-		
+
 		return true;
 	}
 
@@ -564,7 +573,7 @@ class TemplatesModelStyle extends JModelAdmin
 
 		// Clean the cache.
 		$this->cleanCache();
-		
+
 		return true;
 	}
 
@@ -578,7 +587,7 @@ class TemplatesModelStyle extends JModelAdmin
 	{
 		return (object) array('key' => $this->helpKey, 'url' => $this->helpURL);
 	}
-	
+
 	/**
 	 * Custom clean cache method
 	 *
