+++ libraries/joomla/cache/storage/file.php
@@ -1,34 +1,36 @@
 <?php
 /**
- * @version		$Id: file.php 20228 2011-01-10 00:52:54Z eddieajau $
- * @package		Joomla.Framework
- * @subpackage	Cache
- * @copyright	Copyright (C) 2005 - 2011 Open Source Matters, Inc. All rights reserved.
- * @license		GNU General Public License version 2 or later; see LICENSE.txt
+ * @package     Joomla.Platform
+ * @subpackage  Cache
+ *
+ * @copyright   Copyright (C) 2005 - 2011 Open Source Matters, Inc. All rights reserved.
+ * @license     GNU General Public License version 2 or later; see LICENSE
  */
 
-// No direct access
-defined('JPATH_BASE') or die;
+defined('JPATH_PLATFORM') or die;
+
+jimport('joomla.filesystem.file');
 
 /**
  * File cache storage handler
  *
- * @package		Joomla.Framework
- * @subpackage	Cache
- * @since		1.5
+ * @package     Joomla.Platform
+ * @subpackage  Cache
+ * @since       11.1
  */
 class JCacheStorageFile extends JCacheStorage
 {
 	/**
-	 * @since	1.6
+	 * @var    string
+	 * @since  11.1
 	 */
-	private $_root;
+	protected $_root;
 
 	/**
 	 * Constructor
 	 *
-	 * @param	array	$options optional parameters
-	 * @since	1.5
+	 * @param   array    $options optional parameters
+	 * @since   11.1
 	 */
 	public function __construct($options = array())
 	{
@@ -41,11 +43,13 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Get cached data from a file by id and group
 	 *
-	 * @param	string	$id			The cache data id
-	 * @param	string	$group		The cache data group
-	 * @param	boolean	$checkTime	True to verify cache time expiration threshold
-	 * @return	mixed	Boolean false on failure or a cached data string
-	 * @since	1.5
+	 * @param   string   $id         The cache data id
+	 * @param   string   $group      The cache data group
+	 * @param   boolean  $checkTime  True to verify cache time expiration threshold
+	 *
+	 * @return  mixed  Boolean false on failure or a cached data string
+	 *
+	 * @since   11.1
 	 */
 	public function get($id, $group, $checkTime = true)
 	{
@@ -71,8 +75,9 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Get all cached data
 	 *
-	 * @return	array data
-	 * @since	1.6
+	 * @return  array  The cached data
+	 *
+	 * @since   11.1
 	 */
 	public function getAll()
 	{
@@ -84,11 +89,11 @@ class JCacheStorageFile extends JCacheStorage
 
 		foreach ($folders as $folder) {
 			$files 	= array();
-			$files 	= $this->_filesInFolder($path.DS.$folder);
+			$files 	= $this->_filesInFolder($path . '/' . $folder);
 			$item 	= new JCacheStorageHelper($folder);
 
 			foreach ($files as $file) {
-				$item->updateSize(filesize($path.DS.$folder.DS.$file) / 1024);
+				$item->updateSize(filesize($path . '/' . $folder . '/' . $file) / 1024);
 			}
 			$data[$folder] = $item;
 		}
@@ -99,11 +104,13 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Store the data to a file by id and group
 	 *
-	 * @param	string	$id		The cache data id
-	 * @param	string	$group	The cache data group
-	 * @param	string	$data	The data to store in cache
-	 * @return	boolean	True on success, false otherwise
-	 * @since	1.5
+	 * @param   string  $id     The cache data id
+	 * @param   string  $group  The cache data group
+	 * @param   string  $data   The data to store in cache
+	 *
+	 * @return  boolean  True on success, false otherwise
+	 *
+	 * @since   11.1
 	 */
 	public function store($id, $group, $data)
 	{
@@ -133,10 +140,12 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Remove a cached data file by id and group
 	 *
-	 * @param	string	$id		The cache data id
-	 * @param	string	$group	The cache data group
-	 * @return	boolean	True on success, false otherwise
-	 * @since	1.5
+	 * @param   string  $id     The cache data id
+	 * @param   string  $group  The cache data group
+	 *
+	 * @return  boolean  True on success, false otherwise
+	 *
+	 * @since   11.1
 	 */
 	public function remove($id, $group)
 	{
@@ -150,13 +159,14 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Clean cache for a group given a mode.
 	 *
-	 * group mode		: cleans all cache in the group
-	 * notgroup mode	: cleans all cache not in the group
+	 * @param   string  $group  The cache data group
+	 * @param   string  $mode   The mode for cleaning cache [group|notgroup]
+	 *                          group mode     : cleans all cache in the group
+	 *                          notgroup mode  : cleans all cache not in the group
 	 *
-	 * @param	string	$group	The cache data group
-	 * @param	string	$mode	The mode for cleaning cache [group|notgroup]
-	 * @return	boolean	True on success, false otherwise
-	 * @since	1.5
+	 * @return  boolean  True on success, false otherwise
+	 *
+	 * @since   11.1
 	 */
 	public function clean($group, $mode = null)
 	{
@@ -172,14 +182,14 @@ class JCacheStorageFile extends JCacheStorage
 				$folders = $this->_folders($this->_root);
 				for ($i=0, $n=count($folders); $i<$n; $i++) {
 					if ($folders[$i] != $folder) {
-						$return |= $this->_deleteFolder($this->_root.DS.$folders[$i]);
+						$return |= $this->_deleteFolder($this->_root . '/' . $folders[$i]);
 					}
 				}
 				break;
 			case 'group':
 			default:
-				if (is_dir($this->_root.DS.$folder)) {
-					$return = $this->_deleteFolder($this->_root.DS.$folder);
+				if (is_dir($this->_root . '/' . $folder)) {
+					$return = $this->_deleteFolder($this->_root . '/' . $folder);
 				}
 				break;
 		}
@@ -189,8 +199,8 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Garbage collect expired cache data
 	 *
-	 * @return	boolean	True on success, false otherwise.
-	 * @since	1.5
+	 * @return  boolean  True on success, false otherwise.
+	 * @since   11.1
 	 */
 	public function gc()
 	{
@@ -198,6 +208,10 @@ class JCacheStorageFile extends JCacheStorage
 		// files older than lifeTime get deleted from cache
 		$files = $this->_filesInFolder($this->_root, '', true, true);
 		foreach($files As $file) {
+			if('index.html' == JFile::getName($file)) {
+				// Don't delete index.html files
+				continue;
+			}
 			$time = @filemtime($file);
 			if (($time + $this->_lifetime) < $this->_now || empty($time)) {
 				$result |= @unlink($file);
@@ -209,8 +223,8 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Test to see if the cache storage is available.
 	 *
-	 * @return	boolean	True on success, false otherwise.
-	 * @since	1.5
+	 * @return  boolean  True on success, false otherwise.
+	 * @since   11.1
 	 */
 	public static function test()
 	{
@@ -221,15 +235,17 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Lock cached item
 	 *
-	 * @param	string	$id		The cache data id
-	 * @param	string	$group	The cache data group
-	 * @param	integer	$locktime Cached item max lock time
-	 * @return	boolean	True on success, false otherwise.
-	 * @since	1.6
+	 * @param   string   $id        The cache data id
+	 * @param   string   $group     The cache data group
+	 * @param   integer  $locktime  Cached item max lock time
+	 *
+	 * @return  boolean  True on success, false otherwise.
+	 *
+	 * @since   11.1
 	 */
 	public function lock($id,$group,$locktime)
 	{
-		$returning = new stdClass();
+		$returning = new stdClass;
 		$returning->locklooped = false;
 
 		$looptime 	= $locktime * 10;
@@ -270,10 +286,12 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Unlock cached item
 	 *
-	 * @param	string	$id		The cache data id
-	 * @param	string	$group	The cache data group
-	 * @return	boolean	True on success, false otherwise.
-	 * @since	1.6
+	 * @param   string  $id     The cache data id
+	 * @param   string  $group  The cache data group
+	 *
+	 * @return  boolean  True on success, false otherwise.
+	 *
+	 * @since   11.1
 	 */
 	public function unlock($id, $group = null)
 	{
@@ -289,16 +307,17 @@ class JCacheStorageFile extends JCacheStorage
 		return $ret;
 	}
 
-
-
 	/**
 	 * Check to make sure cache is still valid, if not, delete it.
 	 *
-	 * @param	string	$id		Cache key to expire.
-	 * @param	string	$group	The cache data group.
-	 * @since	1.6
+	 * @param   string  $id     Cache key to expire.
+	 * @param   string  $group  The cache data group.
+	 *
+	 * @return  boolean  False if not valid
+	 *
+	 * @since   11.1
 	 */
-	private function _checkExpire($id, $group)
+	protected function _checkExpire($id, $group)
 	{
 		$path = $this->_getFilePath($id, $group);
 
@@ -317,39 +336,43 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Get a cache file path from an id/group pair
 	 *
-	 * @param	string	$id		The cache data id
-	 * @param	string	$group	The cache data group
-	 * @return	string	The cache file path
-	 * @since	1.5
+	 * @param   string  $id     The cache data id
+	 * @param   string  $group  The cache data group
+	 *
+	 * @return  string   The cache file path
+	 *
+	 * @since   11.1
 	 */
-	private function _getFilePath($id, $group)
+	protected function _getFilePath($id, $group)
 	{
 		$name	= $this->_getCacheId($id, $group);
-		$dir	= $this->_root.DS.$group;
+		$dir	= $this->_root . '/' . $group;
 
 		// If the folder doesn't exist try to create it
 		if (!is_dir($dir)) {
 
 			// Make sure the index file is there
 			$indexFile = $dir.'/index.html';
-			@ mkdir($dir) && file_put_contents($indexFile, '<html><body bgcolor="#FFFFFF"></body></html>');
+			@ mkdir($dir) && file_put_contents($indexFile, '<!DOCTYPE html><title></title>');
 		}
 
 		// Make sure the folder exists
 		if (!is_dir($dir)) {
 			return false;
 		}
-		return $dir.DS.$name.'.php';
+		return $dir . '/' . $name.'.php';
 	}
 
 	/**
 	 * Quickly delete a folder of files
 	 *
-	 * @param string The path to the folder to delete.
-	 * @return boolean True on success.
-	 * @since 1.6
+	 * @param   string  The path to the folder to delete.
+	 *
+	 * @return  boolean  True on success.
+	 *
+	 * @since   11.1
 	 */
-	private function _deleteFolder($path)
+	protected function _deleteFolder($path)
 	{
 	// Sanity check
 		if (!$path || !is_dir($path) || empty($this->_root)) {
@@ -367,7 +390,6 @@ class JCacheStorageFile extends JCacheStorage
 			JError::raiseWarning(500, 'JCacheStorageFile::_deleteFolder' . JText::sprintf('JLIB_FILESYSTEM_ERROR_PATH_IS_NOT_A_FOLDER', $path));
 			return false;
 		}
-
 
 		// Remove all the files in folder if they exist; disable all filtering
 		$files = $this->_filesInFolder($path, '.', false, true, array(), array());
@@ -394,7 +416,6 @@ class JCacheStorageFile extends JCacheStorage
 			}
 		}
 
-
 		// Remove sub-folders of folder; disable all filtering
 		$folders = $this->_folders($path, '.', false, true, array(), array());
 
@@ -409,7 +430,6 @@ class JCacheStorageFile extends JCacheStorage
 			}
 		}
 
-
 		// In case of restricted permissions we zap it one way or the other
 		// as long as the owner is either the webserver or the ftp
 		if (@rmdir($path)) {
@@ -421,16 +441,17 @@ class JCacheStorageFile extends JCacheStorage
 		return $ret;
 	}
 
-
 	/**
 	 * Function to strip additional / or \ in a path name
 	 *
-	 * @param	string	The path to clean
-	 * @param	string	Directory separator (optional)
-	 * @return	string	The cleaned path
-	 * @since	1.6
+	 * @param   string  The path to clean
+	 * @param   string  Directory separator (optional)
+	 *
+	 * @return  string  The cleaned path
+	 *
+	 * @since   11.1
 	 */
-	private function _cleanPath($path, $ds = DS)
+	protected function _cleanPath($path, $ds = DS)
 	{
 		$path = trim($path);
 
@@ -448,17 +469,19 @@ class JCacheStorageFile extends JCacheStorage
 	/**
 	 * Utility function to quickly read the files in a folder.
 	 *
-	 * @param	string	The path of the folder to read.
-	 * @param	string	A filter for file names.
-	 * @param	mixed	True to recursively search into sub-folders, or an
-	 * integer to specify the maximum depth.
-	 * @param	boolean	True to return the full path to the file.
-	 * @param	array	Array with names of files which should not be shown in
-	 * the result.
-	 * @return	array	Files in the given folder.
-	 * @since 1.6
+	 * @param   string  The path of the folder to read.
+	 * @param   string  A filter for file names.
+	 * @param   mixed   True to recursively search into sub-folders, or an
+	 *                  integer to specify the maximum depth.
+	 * @param   boolean True to return the full path to the file.
+	 * @param   array   Array with names of files which should not be shown in
+	 *                  the result.
+	 *
+	 * @return  array  Files in the given folder.
+	 *
+	 * @since   11.1
 	 */
-	private function _filesInFolder($path, $filter = '.', $recurse = false, $fullpath = false, $exclude = array('.svn', 'CVS','.DS_Store','__MACOSX'), $excludefilter = array('^\..*','.*~'))
+	protected function _filesInFolder($path, $filter = '.', $recurse = false, $fullpath = false, $exclude = array('.svn', 'CVS','.DS_Store','__MACOSX'), $excludefilter = array('^\..*','.*~'))
 	{
 		// Initialise variables.
 		$arr = array();
@@ -472,7 +495,7 @@ class JCacheStorageFile extends JCacheStorage
 			return false;
 		}
 
-		// read the source directory
+		// Read the source directory.
 		$handle = opendir($path);
 		if (count($excludefilter)) {
 			$excludefilter = '/('. implode('|', $excludefilter) .')/';
@@ -482,7 +505,7 @@ class JCacheStorageFile extends JCacheStorage
 		while (($file = readdir($handle)) !== false)
 		{
 			if (($file != '.') && ($file != '..') && (!in_array($file, $exclude)) && (!$excludefilter || !preg_match($excludefilter, $file))) {
-				$dir = $path . DS . $file;
+				$dir = $path . '/' . $file;
 				$isDir = is_dir($dir);
 				if ($isDir) {
 					if ($recurse) {
@@ -497,7 +520,7 @@ class JCacheStorageFile extends JCacheStorage
 				} else {
 					if (preg_match("/$filter/", $file)) {
 						if ($fullpath) {
-							$arr[] = $path . DS . $file;
+							$arr[] = $path . '/' . $file;
 						} else {
 							$arr[] = $file;
 						}
@@ -510,22 +533,24 @@ class JCacheStorageFile extends JCacheStorage
 		return $arr;
 	}
 
-/**
+	/**
 	 * Utility function to read the folders in a folder.
 	 *
-	 * @param	string	The path of the folder to read.
-	 * @param	string	A filter for folder names.
-	 * @param	mixed	True to recursively search into sub-folders, or an
-	 * integer to specify the maximum depth.
-	 * @param	boolean	True to return the full path to the folders.
-	 * @param	array	Array with names of folders which should not be shown in
-	 * the result.
-	 * @param	array	Array with regular expressions matching folders which
-	 * should not be shown in the result.
-	 * @return	array	Folders in the given folder.
-	 * @since 1.6
+	 * @param   string   The path of the folder to read.
+	 * @param   string   A filter for folder names.
+	 * @param   mixed    True to recursively search into sub-folders, or an
+	 *                   integer to specify the maximum depth.
+	 * @param   boolean  True to return the full path to the folders.
+	 * @param   array    Array with names of folders which should not be shown in
+	 *                   the result.
+	 * @param   array    Array with regular expressions matching folders which
+	 *                   should not be shown in the result.
+	 *
+	 * @return  array  Folders in the given folder.
+	 *
+	 * @since   11.1
 	 */
-	private function _folders($path, $filter = '.', $recurse = false, $fullpath = false, $exclude = array('.svn', 'CVS','.DS_Store','__MACOSX'), $excludefilter = array('^\..*'))
+	protected function _folders($path, $filter = '.', $recurse = false, $fullpath = false, $exclude = array('.svn', 'CVS','.DS_Store','__MACOSX'), $excludefilter = array('^\..*'))
 	{
 		// Initialise variables.
 		$arr = array();
@@ -550,7 +575,7 @@ class JCacheStorageFile extends JCacheStorage
 		while (($file = readdir($handle)) !== false)
 		{
 			if (($file != '.') && ($file != '..') && (!in_array($file, $exclude)) && (empty($excludefilter_string) || !preg_match($excludefilter_string, $file))) {
-				$dir = $path . DS . $file;
+				$dir = $path . '/' . $file;
 				$isDir = is_dir($dir);
 				if ($isDir) {
 					// Removes filtered directories
